Allow opt-out from gopacket via build tag
